+++ NPCs/Boss/AncientHealingMachine/AncientLaser.cs
@@ -53,7 +53,7 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 				Vector2 bsSpawning = new Vector2(-75f, 32f);
 				Vector2 bsSpawningBoo = new Vector2(Main.npc[(int)projectile.ai[1]].Center.X - 75f, Main.npc[(int)projectile.ai[1]].Center.Y + 32f);
 				Vector2 bsSpawningSucks = Utils.Vector2FromElipse(Main.npc[(int)projectile.ai[1]].localAI[0].ToRotationVector2(), bsSpawning * Main.npc[(int)projectile.ai[1]].localAI[1]);
-				projectile.position = bsSpawningBoo + bsSpawningSucks - new Vector2((float)projectile.width, (float)projectile.height) / 2f;
+				projectile.position = bsSpawningBoo + bsSpawningSucks - new Vector2(projectile.width, projectile.height) / 2f;
 			}
 			
 			if (projectile.velocity.HasNaNs() || projectile.velocity == Vector2.Zero)
@@ -71,7 +71,7 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 				projectile.Kill();
 				return;
 			}
-			projectile.scale = (float)Math.Sin((double)(projectile.localAI[0] * 3.141f / 180f)) * 10f * num;
+			projectile.scale = (float)Math.Sin(projectile.localAI[0] * 3.141f / 180f) * 10f * num;
 			if (projectile.scale > num)
 			{
 				projectile.scale = num;
@@ -81,7 +81,7 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 			projectile.rotation = num2 - 1.57f;
 			projectile.velocity = num2.ToRotationVector2();
 			float num3 = 3f;
-			float num4 = (float)projectile.width;
+			float num4 = projectile.width;
 			Vector2 pointWhereItDoesDaDing = projectile.Center;
 			if (vector != null)
 			{
@@ -108,7 +108,7 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 						Main.dust[num813].velocity.Y = 0f - Math.Abs(Main.dust[num813].velocity.Y);
 			}
 			DelegateMethods.v3_1 = new Vector3(0.3f, 0.65f, 0.7f);
-			Utils.PlotTileLine(projectile.Center, projectile.Center + projectile.velocity * projectile.localAI[1], (float)projectile.width * projectile.scale, new Utils.PerLinePoint(DelegateMethods.CastLight));
+			Utils.PlotTileLine(projectile.Center, projectile.Center + projectile.velocity * projectile.localAI[1], projectile.width * projectile.scale, new Utils.PerLinePoint(DelegateMethods.CastLight));
 		}
 
 		public override bool PreDraw(SpriteBatch spriteBatch, Color lightColor)
@@ -143,22 +143,22 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 			Texture2D anotherSpriteBatch = texture2D;
 			Vector2 position = projectile.Center - Main.screenPosition;
 			spriteBatch2.Draw(anotherSpriteBatch, position, null, color, projectile.rotation, texture2D.Size() / 2f, projectile.scale, SpriteEffects.None, 0f);
-			num -= (float)(texture2D.Height / 2 + texture2.Height) * projectile.scale;
+			num -= (texture2D.Height / 2 + texture2.Height) * projectile.scale;
 			Vector2 value = projectile.Center;
-			value += projectile.velocity * projectile.scale * (float)texture2D.Height / 2f;
+			value += projectile.velocity * projectile.scale * texture2D.Height / 2f;
 			if (num > 0f)
 			{
 				float num2 = 0f;
 				Rectangle value2 = new Rectangle(0, 16 * (projectile.timeLeft / 3 % 5), texture.Width, 16);
 				while (num2 + 1f < num)
 				{
-					if (num - num2 < (float)value2.Height)
+					if (num - num2 < value2.Height)
 					{
 						value2.Height = (int)(num - num2);
 					}
-					Main.spriteBatch.Draw(texture, value - Main.screenPosition, new Rectangle?(value2), color, projectile.rotation, new Vector2((float)(value2.Width / 2), 0f), projectile.scale, SpriteEffects.None, 0f);
-					num2 += (float)value2.Height * projectile.scale;
-					value += projectile.velocity * (float)value2.Height * projectile.scale;
+					Main.spriteBatch.Draw(texture, value - Main.screenPosition, new Rectangle?(value2), color, projectile.rotation, new Vector2(value2.Width / 2, 0f), projectile.scale, SpriteEffects.None, 0f);
+					num2 += value2.Height * projectile.scale;
+					value += projectile.velocity * value2.Height * projectile.scale;
 					value2.Y += 16;
 					if (value2.Y + value2.Height > texture.Height)
 					{
@@ -177,7 +177,7 @@ namespace NoxiumMod.NPCs.Boss.AncientHealingMachine
 		{
 			DelegateMethods.tilecut_0 = TileCuttingContext.AttackProjectile;
 			Vector2 velocity = projectile.velocity;
-			Utils.PlotTileLine(projectile.Center, projectile.Center + velocity * projectile.localAI[1], (float)projectile.width * projectile.scale, new Utils.PerLinePoint(DelegateMethods.CutTiles));
+			Utils.PlotTileLine(projectile.Center, projectile.Center + velocity * projectile.localAI[1], projectile.width * projectile.scale, new Utils.PerLinePoint(DelegateMethods.CutTiles));
 		}
 
 		public override bool? Colliding(Rectangle projHitbox, Rectangle targetHitbox)
