+++ Projectiles/BloodyKnifeProjectile.cs
@@ -34,8 +34,8 @@ namespace NoxiumMod.Projectiles
 				timer = 0;
 				if (projectile.owner == Main.myPlayer)
 				{
-					int num421 = (int)(projectile.position.X + 4f + (float)Main.rand.Next(projectile.width - 6));
-					int num422 = (int)(projectile.position.Y + (float)projectile.height + 4f);
+					int num421 = (int)(projectile.position.X + 4f + Main.rand.Next(projectile.width - 6));
+					int num422 = (int)(projectile.position.Y + projectile.height + 4f);
 					Projectile.NewProjectile(num421, num422, 0f, 4f, ProjectileID.BloodRain, projectile.damage, 0f, projectile.owner);
 				}
 			}
